# Conflicts:
#	src/main/java/uz/master/warehouse/services/auth/AuthUserService.java

+++ src/main/java/uz/master/warehouse/services/organization/MarketService.java
@@ -2,9 +2,11 @@ package uz.master.warehouse.services.organization;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import uz.master.warehouse.criteria.GenericCriteria;
 import uz.master.warehouse.dto.market.MarketCreateDto;
 import uz.master.warehouse.dto.market.MarketDto;
 import uz.master.warehouse.dto.market.MarketUpdateDto;
+import uz.master.warehouse.dto.payment.PaymentDto;
 import uz.master.warehouse.dto.responce.AppErrorDto;
 import uz.master.warehouse.dto.responce.DataDto;
 import uz.master.warehouse.entity.organization.Market;
@@ -13,8 +15,9 @@ import uz.master.warehouse.mapper.organization.MarketMapper;
 import uz.master.warehouse.repository.organization.MarketRepository;
 import uz.master.warehouse.services.AbstractService;
 import uz.master.warehouse.services.GenericCrudService;
-import uz.master.warehouse.validator.organization.MarketValidator;
+import uz.master.warehouse.session.SessionUser;
 
+import javax.validation.Valid;
 import java.util.List;
 import java.util.Objects;
 
@@ -27,18 +30,21 @@ public class MarketService extends AbstractService<
         MarketDto,
         MarketCreateDto,
         MarketUpdateDto,
+        GenericCriteria,
         Long
         > {
 
+    private final SessionUser sessionUser;
 
-    public MarketService(MarketRepository repository,  MarketMapper mapper) {
+
+    public MarketService(MarketRepository repository, MarketMapper mapper, SessionUser sessionUser) {
         super(repository, mapper);
+        this.sessionUser = sessionUser;
     }
 
 
     @Override
-    public DataDto<Long> create(MarketCreateDto createDto) {
-
+    public DataDto<Long> create(@Valid MarketCreateDto createDto) {
         Market market = mapper.fromCreateDto(createDto);
         Organization organization = repository.findByOrgId(createDto.getOrganizationId());
         if (Objects.isNull(organization)) {
@@ -55,18 +61,24 @@ public class MarketService extends AbstractService<
 
     @Override
     public DataDto<Void> delete(Long id) {
+        DataDto<MarketDto> marketDtoDataDto = get(id);
+        Long organizationId = marketDtoDataDto.getData().getOrganizationId();
+        Long orgId = sessionUser.getOrgId();
+        if (!organizationId.equals(orgId)) {
+            return new DataDto<>(new AppErrorDto(HttpStatus.FORBIDDEN, "You have no such privilege", "market/delete"));
+        }
         repository.deleteMarket(id);
         return new DataDto<>();
     }
 
     @Override
-    public DataDto<Long> update(MarketUpdateDto updateDto) {
+    public DataDto<Long> update(@Valid MarketUpdateDto updateDto) {
 
         Market market = mapper.fromUpdateDto(updateDto);
         market.setName(updateDto.getName());
         market.setDescription(updateDto.getDescription());
         market.setLocation(updateDto.getLocation());
-          repository.update(market.getId(), market.getName(), market.getLocation(), market.getDescription());
+        repository.update(market.getId(), market.getName(), market.getLocation(), market.getDescription());
         return new DataDto<>(market.getId());
     }
 
@@ -79,10 +91,15 @@ public class MarketService extends AbstractService<
     @Override
     public DataDto<MarketDto> get(Long id) {
         Market market = repository.findByIdAndDeletedFalse(id);
-        if (Objects.isNull(market)){
+        if (Objects.isNull(market)) {
             return new DataDto<>(new AppErrorDto(HttpStatus.NOT_FOUND, "Market not found", "market/get"));
 
         }
         return new DataDto<>(mapper.toDto(market));
+    }
+
+    @Override
+    public DataDto<List<MarketDto>> getWithCriteria(GenericCriteria criteria) {
+        return null;
     }
 }
