+++ src/main/java/uz/master/warehouse/services/payment/PaymentService.java
@@ -1,8 +1,8 @@
 package uz.master.warehouse.services.payment;
 
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import uz.master.warehouse.criteria.PaymentCriteria;
 import uz.master.warehouse.dto.payment.PaymentCreateDto;
 import uz.master.warehouse.dto.payment.PaymentDto;
 import uz.master.warehouse.dto.payment.PaymentUpdateDto;
@@ -13,11 +13,10 @@ import uz.master.warehouse.mapper.payment.PaymentMapper;
 import uz.master.warehouse.repository.payment.PaymentRepository;
 import uz.master.warehouse.services.AbstractService;
 import uz.master.warehouse.services.GenericCrudService;
-import uz.master.warehouse.validator.payment.PaymentValidator;
+import uz.master.warehouse.session.SessionUser;
 
 import javax.validation.Valid;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
@@ -29,17 +28,19 @@ public class PaymentService extends AbstractService<
         PaymentDto,
         PaymentCreateDto,
         PaymentUpdateDto,
+        PaymentCriteria,
         Long> {
 
 
-    public PaymentService(PaymentRepository repository,
-                          PaymentMapper mapper) {
+    private final SessionUser sessionUser;
+
+
+    public PaymentService(PaymentRepository repository, PaymentMapper mapper, SessionUser sessionUser) {
         super(repository, mapper);
+        this.sessionUser = sessionUser;
     }
 
-    @Override
     public DataDto<Long> create(@Valid PaymentCreateDto createDto) {
-
         Payment payment = mapper.fromCreateDto(createDto);
         payment.setOrganizationId(createDto.getOrganizationId());
         payment.setCompanyId(createDto.getCompanyId());
@@ -50,13 +51,17 @@ public class PaymentService extends AbstractService<
 
     @Override
     public DataDto<Void> delete(Long id) {
+        DataDto<PaymentDto> paymentDtoDataDto = get(id);
+        Long organizationId = paymentDtoDataDto.getData().getOrganizationId();
+        Long orgId = sessionUser.getOrgId();
+        if (!organizationId.equals(orgId)) {
+            return new DataDto<>(new AppErrorDto(HttpStatus.FORBIDDEN, "You have no such privilege", "payment/delete"));
+        }
         repository.deletePayment(id);
         return new DataDto<>();
     }
 
-    @Override
-    public DataDto<Long> update(PaymentUpdateDto updateDto) {
-
+    public DataDto<Long> update(@Valid PaymentUpdateDto updateDto) {
         Payment payment = mapper.fromUpdateDto(updateDto);
         payment.setSum(updateDto.getSum());
         repository.updatePayment(payment.getId(), payment.getSum());
@@ -73,10 +78,15 @@ public class PaymentService extends AbstractService<
     public DataDto<PaymentDto> get(Long id) {
         Payment payment = repository.findByIdAndDeletedFalse(id);
         if (Objects.isNull(payment)) {
-            return new DataDto<>(new AppErrorDto(HttpStatus.NOT_FOUND, "Payment not found", "payment/get"));
-
+            return new DataDto<>(
+                    new AppErrorDto(HttpStatus.NOT_FOUND, "Payment not found", "payment/get"));
         }
         return new DataDto<>(mapper.toDto(payment));
+    }
+
+    @Override
+    public DataDto<List<PaymentDto>> getWithCriteria(PaymentCriteria criteria) {
+        return null;
     }
 
     public DataDto<List<PaymentDto>> getByTime(String fromDate, String toDate) {
@@ -84,5 +94,12 @@ public class PaymentService extends AbstractService<
         LocalDate to = LocalDate.parse(toDate);
         List<Payment> allByDateTimeDateBetween = repository.findAllByDateTimeDateBetween(from, to);
         return new DataDto<>(mapper.toDto(allByDateTimeDateBetween));
+    }
+
+    public List<PaymentDto> getByTimeBetween(String fromDate, String toDate) {
+        LocalDate from = LocalDate.parse(fromDate);
+        LocalDate to = LocalDate.parse(toDate);
+        List<Payment> allByDateTimeDateBetween = repository.findAllByDateTimeDateBetween(from, to);
+        return mapper.toDto(allByDateTimeDateBetween);
     }
 }
